requirements.txt
Flask
gunicorn
neo4j
python-dotenv
requests
google-generativeai
reportlab
